src/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
geopandas==0.14.3
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0
networkx==3.2.1
shapely==2.0.3
python-multipart==0.0.9
rasterio>=1.3.0
numba>=0.57.0
cupy-cuda12x